% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boofactors-package.r
\docType{package}
\name{boofactors}
\alias{boofactors}
\alias{boofactors-package}
\title{boofactors: Make factors less annoying}
\description{
Factors have driven people to extreme measures, like ordering custom
conference ribbons and laptop stickers to express how HELLNO we feel about
stringsAsFactors. And yet, sometimes you need them. Can they be made less
maddening? Let's find out.
}
